hibiki/app/requirements.txt
apscheduler==3.2.0
beautifulsoup4==4.4.1
bottle==0.12.9
fluent-logger==0.4.2
google-api-python-client==1.5.0
html5lib==0.9999999
httplib2shim==0.0.1
jinja2==2.8
matplotlib==1.5.1
nose==1.3.7
oauth2client==2.2.0
passlib==1.6.5
paste==2.0.3
pylint==1.5.6
pymongo==3.2.2
python-gflags==3.0.4
requests==2.10.0
subprocess32==3.2.7
ujson==1.35
uwsgi==2.0.12
wsgiprof==1.0.0
